WIP: Windows virtual IP notes
When not using skipAsSource with the installed virtual IP, the IP gets promoted
as source address for the already available routes over that interface. When
setting the flag, all of our manually installed IPsec routes still use
the interfaces main address as source address, as the route uses that interface.

To fix this issue, we probably need a dedicated interface for virtual IPs that
allows us to install our separated routes over that interface.

Using the MS Loopback adapter kinda works; when disabling skipAsSource, an
address installed to that adapter gets used and outgoing traffic flows as
expected. Inbound traffic, though, fails with STATUS_IPSEC_CLEAR_TEXT_DROP,
probably related to:

https://wiki.strongswan.org/projects/strongswan/wiki/Kernel-wfp#Accessing-Gateway-internal-address-in-a-net-to-net-tunnel